Add the StandaloneUnfolding class. Also restructure things so that a lot of code originally defined in header files now lives in source files. This avoids linker problems due to duplicate definitions as we add new source files.
